requirements.txt
aiogram==3.4.1
asyncpg
SQLAlchemy==2.0.25
python-dotenv
